chore(refactor): move classes Util.AbsoluteDirection -> AbsoluteDirection, Util.RelativeDirection -> RelativeDirection
